+++ cmd/go-camo/main.go
@@ -51,6 +51,7 @@ func main() {
 		Verbose             bool          `short:"v" long:"verbose" description:"Show verbose (debug) log level output"`
 		ServerName          string        `long:"server-name" default:"go-camo" description:"Value to use for the HTTP server field"`
 		ExposeServerVersion bool          `long:"expose-server-version" description:"Include the server version in the HTTP server response header"`
+		EnableXFwdFor       bool          `long:"enable-xfwd4" description:"Enable x-forward-for passthrough/generation"`
 	}
 
 	// parse said flags
@@ -115,6 +116,9 @@ func main() {
 	// set keepalive options
 	config.DisableKeepAlivesBE = opts.DisableKeepAlivesBE
 	config.DisableKeepAlivesFE = opts.DisableKeepAlivesFE
+
+	// other options
+	config.EnableXFwdFor = opts.EnableXFwdFor
 
 	// additonal content types to allow
 	config.AllowContentVideo = opts.AllowContentVideo